+++ strings/country_code.go
@@ -1,9 +1,0 @@
-package strings
-
-type CountryCode struct {
-	lower
-}
-
-func NewCountryCode(code string) CountryCode {
-	return CountryCode{newLower(code)}
-}