+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Release Notes
 
-## v1.0.0 / 2020-07-28
+## v1.0.0 / 2020-07-31
 - project created
+- internal/Lower base type added
+- CountryCode, Language, Nationality added